+++ starsim/interventions.py
@@ -7,7 +7,6 @@ import starsim as ss
 import sciris as sc
 import numpy as np
 
-__all__ = ['Intervention']
 
 
 class Intervention(ss.Module):
@@ -69,7 +68,7 @@ class Intervention(ss.Module):
 
 # %% Template classes for routine and campaign delivery
 
-__all__ += ['RoutineDelivery', 'CampaignDelivery']
+
 
 class RoutineDelivery(Intervention):
     """
@@ -170,8 +169,7 @@ class CampaignDelivery(Intervention):
 
 # %% Screening and triage
 
-__all__ += ['BaseTest', 'BaseScreening', 'routine_screening', 'campaign_screening', 'BaseTriage', 'routine_triage',
-            'campaign_triage']
+
 
 class BaseTest(Intervention):
     """
@@ -341,7 +339,7 @@ class campaign_triage(BaseTriage, CampaignDelivery):
 
 #%% Treatment interventions
 
-__all__ += ['BaseTreatment', 'treat_num']
+
 
 class BaseTreatment(Intervention):
     """
@@ -442,7 +440,7 @@ class treat_num(BaseTreatment):
 
 #%% Vaccination
 
-__all__ += ['BaseVaccination', 'routine_vx', 'campaign_vx']
+
 
 class BaseVaccination(Intervention):
     """