+++ starsim/parameters.py
@@ -8,7 +8,6 @@ import sciris as sc
 import starsim as ss
 import datetime as dt
 
-__all__ = ['Pars', 'SimPars']
 
 # Define classes to not descend into further -- based on sciris.sc_nested
 atomic_classes = (str, Number, list, np.ndarray, pd.Series, pd.DataFrame, type(None), dt.date)